+++ src/Resources/contao/languages/en/tl_wertungsportal_referenten.php
@@ -1,0 +1,22 @@
+<?php
+
+/**
+ * Labels for table tl_wertungsportal_referenten (fallback)
+ */
+
+$GLOBALS['TL_LANG']['tl_wertungsportal_referenten']['nachname']  = array('Last name', 'Last name of the rating officer');
+$GLOBALS['TL_LANG']['tl_wertungsportal_referenten']['vorname']   = array('First name', 'First name of the rating officer');
+$GLOBALS['TL_LANG']['tl_wertungsportal_referenten']['nuId']      = array('nu ID', 'Person ID in the rating portal, e.g. NU4093214');
+$GLOBALS['TL_LANG']['tl_wertungsportal_referenten']['email']     = array('E-mail', 'E-mail address');
+$GLOBALS['TL_LANG']['tl_wertungsportal_referenten']['telefon']   = array('Phone', 'Phone number');
+$GLOBALS['TL_LANG']['tl_wertungsportal_referenten']['strasse']   = array('Street', 'Street and house number');
+$GLOBALS['TL_LANG']['tl_wertungsportal_referenten']['plz']       = array('Postcode', 'Postal code');
+$GLOBALS['TL_LANG']['tl_wertungsportal_referenten']['ort']       = array('City', 'City');
+$GLOBALS['TL_LANG']['tl_wertungsportal_referenten']['verbaende'] = array('Responsible for', 'Associations this officer is responsible for. The list is read from the club records.');
+$GLOBALS['TL_LANG']['tl_wertungsportal_referenten']['published'] = array('Publish', 'Only published officers appear in the front end');
+
+$GLOBALS['TL_LANG']['tl_wertungsportal_referenten']['person_legend']    = 'Person';
+$GLOBALS['TL_LANG']['tl_wertungsportal_referenten']['kontakt_legend']   = 'Contact';
+$GLOBALS['TL_LANG']['tl_wertungsportal_referenten']['adresse_legend']   = 'Address';
+$GLOBALS['TL_LANG']['tl_wertungsportal_referenten']['verband_legend']   = 'Responsibility';
+$GLOBALS['TL_LANG']['tl_wertungsportal_referenten']['published_legend'] = 'Publication';